requirements.txt
pillow==10.4.0
pdfix-sdk==8.7.6
requests==2.32.4
torch==2.8.0
transformers==4.56.2
types-requests==2.32.0.20250515
